feat: Moiving <form> out of Main and splitting in Buy and  Sell forms. Buy and Sell button are meant to swap forms but functionality is still not there, just the UI
